@@ -3,13 +3,14 @@
   "version": "1.0.1",
   "description": "Titlecase converter for french",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "./node_modules/.bin/mocha --reporter spec",
     "lint": "eslint .; exit 0"
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/benoitvallon/titlecase-french.git"
+    "url": "git://github.com/ChrisJohnNewton/titlecase-french.git"
   },
   "keywords": [
     "titlecase",
@@ -21,8 +22,8 @@
   "author": "Benoit VALLON <benoitvallon@gmail.com> (http://blog.benoitvallon.com)",
   "license": "MIT",
   "devDependencies": {
-    "chai": "^3.4.0",
-    "eslint": "^1.9.0",
-    "mocha": "^2.3.3"
+    "chai": "^5.1.1",
+    "eslint": "^9.11.1",
+    "mocha": "^10.7.3"
   }
 }
